requirements/apps.txt
django
django-piston
pygments
chardet
pyyaml
cheetah
